Fix potential int overflow on 16-bit system

+++ Fusion/FusionAhrs.c
@@ -33,7 +33,7 @@ static inline FusionVector HalfMagnetic(const FusionAhrs *const ahrs);
 
 static inline FusionVector Feedback(const FusionVector sensor, const FusionVector reference);
 
-static inline int Clamp(const int value, const int min, const int max);
+static inline int32_t Clamp(const int32_t value, const int32_t min, const int32_t max);
 
 //------------------------------------------------------------------------------
 // Variables
@@ -75,10 +75,10 @@ void FusionAhrsRestart(FusionAhrs *const ahrs) {
     ahrs->halfMagnetometerFeedback = FUSION_VECTOR_ZERO;
     ahrs->accelerometerIgnored = false;
     ahrs->accelerationRecoveryTrigger = 0;
-    ahrs->accelerationRecoveryTimeout = ahrs->settings.recoveryTriggerPeriod;
+    ahrs->accelerationRecoveryTimeout = (int32_t) ahrs->settings.recoveryTriggerPeriod;
     ahrs->magnetometerIgnored = false;
     ahrs->magneticRecoveryTrigger = 0;
-    ahrs->magneticRecoveryTimeout = ahrs->settings.recoveryTriggerPeriod;
+    ahrs->magneticRecoveryTimeout = (int32_t) ahrs->settings.recoveryTriggerPeriod;
 }
 
 /**
@@ -94,8 +94,8 @@ void FusionAhrsSetSettings(FusionAhrs *const ahrs, const FusionAhrsSettings *con
     ahrs->settings.accelerationRejection = settings->accelerationRejection == 0.0f ? FLT_MAX : powf(0.5f * sinf(FusionDegreesToRadians(settings->accelerationRejection)), 2);
     ahrs->settings.magneticRejection = settings->magneticRejection == 0.0f ? FLT_MAX : powf(0.5f * sinf(FusionDegreesToRadians(settings->magneticRejection)), 2);
     ahrs->settings.recoveryTriggerPeriod = settings->recoveryTriggerPeriod;
-    ahrs->accelerationRecoveryTimeout = ahrs->settings.recoveryTriggerPeriod;
-    ahrs->magneticRecoveryTimeout = ahrs->settings.recoveryTriggerPeriod;
+    ahrs->accelerationRecoveryTimeout = (int32_t) ahrs->settings.recoveryTriggerPeriod;
+    ahrs->magneticRecoveryTimeout = (int32_t) ahrs->settings.recoveryTriggerPeriod;
     if ((settings->gain == 0.0f) || (settings->recoveryTriggerPeriod == 0)) {
         ahrs->settings.accelerationRejection = FLT_MAX; // disable acceleration and magnetic rejection features if gain is zero
         ahrs->settings.magneticRejection = FLT_MAX;
@@ -168,9 +168,9 @@ void FusionAhrsUpdate(FusionAhrs *const ahrs, const FusionVector gyroscope, cons
             ahrs->accelerationRecoveryTimeout = 0;
             ahrs->accelerometerIgnored = false;
         } else {
-            ahrs->accelerationRecoveryTimeout = ahrs->settings.recoveryTriggerPeriod;
+            ahrs->accelerationRecoveryTimeout = (int32_t) ahrs->settings.recoveryTriggerPeriod;
         }
-        ahrs->accelerationRecoveryTrigger = Clamp(ahrs->accelerationRecoveryTrigger, 0, ahrs->settings.recoveryTriggerPeriod);
+        ahrs->accelerationRecoveryTrigger = Clamp(ahrs->accelerationRecoveryTrigger, 0, (int32_t) ahrs->settings.recoveryTriggerPeriod);
 
         // Apply accelerometer feedback
         if (ahrs->accelerometerIgnored == false) {
@@ -201,9 +201,9 @@ void FusionAhrsUpdate(FusionAhrs *const ahrs, const FusionVector gyroscope, cons
             ahrs->magneticRecoveryTimeout = 0;
             ahrs->magnetometerIgnored = false;
         } else {
-            ahrs->magneticRecoveryTimeout = ahrs->settings.recoveryTriggerPeriod;
+            ahrs->magneticRecoveryTimeout = (int32_t) ahrs->settings.recoveryTriggerPeriod;
         }
-        ahrs->magneticRecoveryTrigger = Clamp(ahrs->magneticRecoveryTrigger, 0, ahrs->settings.recoveryTriggerPeriod);
+        ahrs->magneticRecoveryTrigger = Clamp(ahrs->magneticRecoveryTrigger, 0, (int32_t) ahrs->settings.recoveryTriggerPeriod);
 
         // Apply magnetometer feedback
         if (ahrs->magnetometerIgnored == false) {
@@ -321,7 +321,7 @@ static inline FusionVector Feedback(const FusionVector sensor, const FusionVecto
  * @param max Maximum value.
  * @return Value limited to maximum and minimum.
  */
-static inline int Clamp(const int value, const int min, const int max) {
+static inline int32_t Clamp(const int32_t value, const int32_t min, const int32_t max) {
     if (value < min) {
         return min;
     }
